Reorganize code on new files/subfolders and better naming schemes
